Adding multiple instances support
