+++ src/utilities/option-utilities-factory.js
@@ -1,0 +1,66 @@
+/*
+ *  Extend default options and get current options by using this file
+ */
+
+module.exports = function () {
+  // default options
+  var defaults = {
+    // The path of core library images when sbgnviz is required from npm and the index html
+    // file and node_modules are under the same folder then using the default value is fine
+    imgPath: 'node_modules/sbgnviz/src/img',
+    // Whether to fit labels to nodes
+    fitLabelsToNodes: function () {
+      return false;
+    },
+    fitLabelsToInfoboxes: function () {
+      return false;
+    },
+    // dynamic label size it may be 'small', 'regular', 'large'
+    dynamicLabelSize: function () {
+      return 'regular';
+    },
+    // percentage used to calculate compound paddings
+    compoundPadding: function () {
+      return 10;
+    },
+    improveFlow: function () {
+        return true;
+    },
+    // extra padding for compartment and complexes
+    extraCompartmentPadding: 10,
+    extraComplexPadding: 10,
+    // Wether to display the complex's labels, like compartments.
+    // Will also increase the paddings by extraCompoundPadding to make room for the name.
+    showComplexName: false,
+    // The selector of the component containing the sbgn network
+    networkContainerSelector: '#sbgn-network-container',
+    // Whether the actions are undoable, requires cytoscape-undo-redo extension
+    undoable: true
+  };
+
+  var optionUtilities = function (param) {
+  };
+
+  // Extend the defaults options with the user options
+  optionUtilities.extendOptions = function (options) {
+    var result = {};
+
+    for (var prop in defaults) {
+      result[prop] = defaults[prop];
+    }
+
+    for (var prop in options) {
+      result[prop] = options[prop];
+    }
+
+    optionUtilities.options = result;
+
+    return options;
+  };
+
+  optionUtilities.getOptions = function () {
+    return optionUtilities.options;
+  };
+
+  return optionUtilities;
+};
