--- src/utilities/ui-utilities.js
@@ -1,35 +1,0 @@
-/*
- * Commonly needed UI Utilities
- */
-
-var optionUtilities = require('./option-utilities');
-var options = optionUtilities.getOptions();
-var libs = require('./lib-utilities').getLibs();
-var jQuery = $ = libs.jQuery;
-
-var uiUtilities = {
-  startSpinner: function (className) {
-    if (!className) {
-      className = 'default-class';
-    }
-    
-    if ($('.' + className).length === 0) {
-      var containerWidth = $(options.networkContainerSelector).width();
-      var containerHeight = $(options.networkContainerSelector).height();
-      $(options.networkContainerSelector + ':parent').prepend('<i style="position: absolute; z-index: 9999999; left: ' + containerWidth / 2 + 'px; top: ' + containerHeight / 2 + 'px;" class="fa fa-spinner fa-spin fa-3x fa-fw ' + className + '"></i>');
-    }
-  },
-  endSpinner: function (className) {
-    if (!className) {
-      className = 'default-class';
-    }
-    
-    if ($('.' + className).length > 0) {
-      $('.' + className).remove();
-    }
-  }
-};
-
-module.exports = uiUtilities;
-
-
