+++ src/utilities/undo-redo-action-functions-factory.js
@@ -1,24 +1,35 @@
 /*
- * This file exports the functions to be utilized in undoredo extension actions 
+ * This file exports the functions to be utilized in undoredo extension actions
  */
-var elementUtilities = require('./element-utilities');
 
-var undoRedoActionFunctions = {
-  deleteElesSimple: function (param) {
+module.exports = function () {
+
+  var elementUtilities;
+  var cy;
+
+  function undoRedoActionFunctions (param) {
+    elementUtilities = param.elementUtilities;
+    cy = param.sbgnCyInstance.getCy();
+  }
+
+  undoRedoActionFunctions.deleteElesSimple = function (param) {
     return elementUtilities.deleteElesSimple(param.eles);
-  },
-  restoreEles: function (eles) {
+  };
+
+  undoRedoActionFunctions.restoreEles = function (eles) {
     var param = {};
     param.eles = elementUtilities.restoreEles(eles);
     return param;
-  },
-  deleteNodesSmart: function (param) {
+  };
+
+  undoRedoActionFunctions.deleteNodesSmart = function (param) {
     if (param.firstTime) {
       return elementUtilities.deleteNodesSmart(param.eles);
     }
     return elementUtilities.deleteElesSimple(param.eles);
-  },
-  setPortsOrdering: function(param) {
+  };
+
+  undoRedoActionFunctions.setPortsOrdering = function(param) {
     var nodes = param.nodes;
     var ordering = param.ordering;
     var portDistance = param.portDistance;
@@ -77,7 +88,7 @@ var undoRedoActionFunctions = {
     }
 
     return result;
-  }
-};
+  };
 
-module.exports = undoRedoActionFunctions;
+  return undoRedoActionFunctions;
+};
